More Mammon references and some bugfixes
Added Mammon death animation;
Added reference to Mammon stats table (temporary values);
Added temporary animations to all of Mammon's skills;
Fixed Mammon's AI hp calculation using integer values instead of float;
Made death animation non-loopable for all units;
SkillVFX now supports creation of multiple particle effects;
Added a safeguard to make sure animation triggered skills are only used
once per turn (does not affect usage of callback skills);
Fixed values on Surplus skill table;
Fixed meta information for a few sprites;

+++ Assets/Scripts/Scriptable Objects/EnemiesAI/Mammon.cs
@@ -21,7 +21,7 @@ namespace FinalInferno{
             float rand = Random.Range(0.0f, 1.0f); //gera um numero aleatorio entre 0 e 1
             float percentageDebuff = Mathf.Min((1f/3f), percentageNotDefense/3f); //porcentagem para o inimigo usar a habilidade de debuff
             BattleUnit battleUnit = BattleManager.instance.GetTeam(UnitType.Enemy)[0];
-            float hpPercent = battleUnit.CurHP / battleUnit.MaxHP;
+            float hpPercent = battleUnit.CurHP / (float)battleUnit.MaxHP;
 
             // Se o hp passar do threshold, usa a skill de regen
             // habilidade pode ser usada 3 vezes, cada uso move o threshold
